// src/models/survey-elements/scoreTable.ts
export interface ScoreTable {
    sId: number;
    elementId: number;
    elementTitle: string;
    situation: string;
    questiontext: string;
    criterion1: string;
    criterion2: string;
    criterion3: string;
    criterion4: string;
    criterion5: string;
    criterion6: string;
    picture: string;
    surveyId: number;
}